Move things out of index.js
