+++ index.js
@@ -36,412 +36,21 @@ nunjucksEnv.addFilter('isString', val => typeof val === 'string');
  */
 const extras = require('sassdoc-extras');
 
-const byGroup = (data, orderedGroups) => {
-  const dataByGroup = {};
-  data.forEach(item => {
-    const group = item.group[0];
-    if (!dataByGroup.hasOwnProperty(group)) {
-      dataByGroup[group] = [];
-    }
-    dataByGroup[group].push(item);
-  });
-  // Sort results by order groups were defined
-  const sorted = {};
-  for (const group of Object.keys(orderedGroups)) {
-    if (dataByGroup.hasOwnProperty(group)) {
-      sorted[group] = dataByGroup[group];
-    }
-  }
-  return sorted;
-};
+const byGroup = require('./lib/byGroup');
 
-const prepareContext = ctx => {
-  const def = {
-    display: {
-      access: ['public', 'private'],
-      alias: false,
-      watermark: true,
-    },
-    groups: {
-      undefined: 'General',
-    },
-    sort: ['group', 'file', 'line', 'access'],
-    herman: { sass: {} },
-  };
+const prepareContext = require('./lib/prepareContext');
 
-  // Apply default values for groups and display.
-  ctx.groups = extend(def.groups, ctx.groups);
-  ctx.display = extend(def.display, ctx.display);
-
-  if (!ctx.description) {
-    def.descriptionPath = './README.md';
-  }
-
-  // Extend top-level context keys.
-  ctx = extend({}, def, ctx);
-
-  /**
-   * Load a `sass-json file` (if one is given in the context) and add its
-   * contents under the `sassjson` key of the context.
-   */
-  if (ctx.herman.sass && ctx.herman.sass.jsonfile && !ctx.sassjson) {
-    ctx.sassjson = parse.sassJson(fs.readFileSync(ctx.herman.sass.jsonfile));
-  }
-
-  /**
-   * Remove the bogus context from any standalone sassdoc comments.
-   * (We detect these if the context starts more than one line after the
-   * sassdoc comment ends.)
-   */
-  ctx.data.forEach(item => {
-    if (!item.context || !item.context.line) {
-      return;
-    }
-    if (item.context.type === 'unknown') {
-      item.context.type = 'prose';
-      item.context.line.end = item.context.line.start;
-    }
-    // Consider it to be prose if it's separated from the next Sass block
-    // by any blank lines.
-    const name = item.context.origName || item.context.name || '';
-    const lineCount = name.split('\n').length;
-    if (!name || item.context.line.start > item.commentRange.end + lineCount) {
-      item.context = {
-        type: 'prose',
-        line: item.commentRange,
-      };
-    }
-  });
-
-  /**
-   * Add `description` and `descriptionPath` from configuration.
-   * Note: `descriptionPath` will override `description`.
-   *
-   * See
-   * <http://sassdoc.com/extra-tools/#description-description-descriptionpath>.
-   */
-  extras.description(ctx);
-
-  /**
-   * Parse text data (like descriptions) as Markdown, and put the
-   * rendered HTML in `html*` variables.
-   *
-   * For example, `ctx.package.description` will be parsed as Markdown
-   * in `ctx.package.htmlDescription`.
-   *
-   * See <http://sassdoc.com/extra-tools/#markdown-markdown>.
-   */
-  extras.markdown(ctx);
-
-  /**
-   * Add a `display` property for each data item regarding of display
-   * configuration (hide private items and aliases for example).
-   *
-   * You'll need to add default values in your `.sassdocrc` before
-   * using this filter:
-   *
-   *     {
-   *       "display": {
-   *         "access": ["public", "private"],
-   *         "alias": false
-   *       }
-   *     }
-   *
-   * See <http://sassdoc.com/extra-tools/#display-toggle-display>.
-   */
-  extras.display(ctx);
-
-  /**
-   * Allow the user to give a name to the documentation groups.
-   *
-   * We can then have `@group slug` in the docblock, and map `slug`
-   * to `Some title string` in the theme configuration.
-   *
-   * **Note:** all items without a group are in the `undefined` group.
-   *
-   * See <http://sassdoc.com/extra-tools/#groups-aliases-groupname>.
-   */
-  extras.groupName(ctx);
-
-  /**
-   * Converts `shortcutIcon` config option into an object:
-   *
-   *       {
-   *         "type": "external",
-   *         "url": "http://absolute.path/to/icon.png"
-   *       }
-   *
-   *       {
-   *         "type": "internal",
-   *         "url": "icon.png",
-   *         "path": "/complete/relative/path/to/icon.png"
-   *       }
-   *
-   * See <http://sassdoc.com/extra-tools/#shortcut-icon-shortcuticon>.
-   */
-  extras.shortcutIcon(ctx);
-
-  /**
-   * Sorts items based on the `sort` config value.
-   * Sort order is determined by the last character: > (desc) or < (asc).
-   *
-   *       {
-   *         "sort": [
-   *            "access",
-   *            "line>",
-   *            "group",
-   *            "file"
-   *          ]
-   *       }
-   *
-   * See <http://sassdoc.com/extra-tools/#sort-sort>.
-   */
-  extras.sort(ctx);
-
-  /**
-   * Connects aliased variables to their original value.
-   * Adds `resolvedValue` key:
-   *
-   *       {
-   *         "description": "<p>Main color</p>\n",
-   *         "context": {
-   *           "type": "variable",
-   *           "name": "color-main",
-   *           "value": "$color-blue"
-   *         },
-   *         "type": "Color",
-   *         "resolvedValue": "#22b8dc"
-   *       }
-   *
-   * See <http://sassdoc.com/extra-tools/#resolved-variables-resolvevariables>.
-   */
-  extras.resolveVariables(ctx);
-
-  /**
-   * Index the data by group, so we have the following structure:
-   *
-   *     {
-   *       "group-slug": [...],
-   *       "another-group": [...]
-   *     }
-   *
-   * You can then use `data.byGroup` instead of `data` in your
-   * templates to manipulate the indexed object.
-   */
-  ctx.byGroup = byGroup(ctx.data, ctx.groups);
-
-  return ctx;
-};
-
-const parseSubprojects = ctx => {
-  const promises = [];
-  if (ctx.herman.subprojects) {
-    ctx.subprojects = {};
-    ctx.herman.subprojects.forEach(name => {
-      const prjPath = `./node_modules/${name}/`;
-      const configFile = `${prjPath}.sassdocrc`;
-      const packageFile = `${prjPath}package.json`;
-      let config = {};
-      let packageJSON = {};
-      try {
-        // Load .sassdocrc configuration from subproject directory
-        config = yaml.safeLoad(fs.readFileSync(configFile, 'utf-8'));
-      } catch (err) {
-        ctx.logger.warn(
-          `Invalid or no .sassdocrc found for subproject: ${name}`
-        );
-      }
-      try {
-        // Load package.json data from subproject directory
-        packageJSON = yaml.safeLoad(fs.readFileSync(packageFile, 'utf-8'));
-      } catch (err) {
-        ctx.logger.warn(
-          `Invalid or no package.json found for subproject: ${name}`
-        );
-      }
-      if (!config.description && !config.descriptionPath) {
-        // Set default descriptionPath for subproject
-        config.descriptionPath = `${prjPath}README.md`;
-      }
-      let src;
-      if (config.src) {
-        // Set subproject src files based on .sassdocrc `src` option
-        src = path.isAbsolute(config.src) ? config.src : prjPath + config.src;
-      } else {
-        // Fall back to all subproject `.scss` files
-        src = `${prjPath}**/*.scss`;
-      }
-      // Remove unused/unnecessary config options
-      config.src = undefined;
-      config.theme = undefined;
-      config.dest = undefined;
-      const prjCtx = extend({}, config);
-      const promise = sassdoc.parse(src, prjCtx).then(data => {
-        prjCtx.subpackage = packageJSON;
-        prjCtx.package = ctx.package;
-        prjCtx.basePath = '../';
-        prjCtx.activeProject = name;
-        prjCtx.data = data;
-        prjCtx.subprojects = ctx.subprojects;
-        prjCtx.topGroups = ctx.groups;
-        prjCtx.topByGroup = ctx.byGroup;
-        ctx.subprojects[name] = prepareContext(prjCtx);
-      });
-      promises.push(promise);
-    });
-  }
-  return Promise.all(promises);
-};
+const parseSubprojects = require('./lib/parseSubprojects');
 
 /**
  * Actual theme function. It takes the destination directory `dest`,
  * and the context variables `ctx`.
  */
-const renderHerman = (dest, ctx) => {
-  const indexTemplate = path.join(base, 'index.j2');
-  const indexDest = path.join(dest, 'index.html');
-  const groupTemplate = path.join(base, 'group.j2');
-  const assets = path.resolve(__dirname, './dist');
-
-  // Accepts a color (in any format) and returns an object with hex, rgba, and
-  // hsla strings.
-  nunjucksEnv.addFilter('colors', input => {
-    const color = tinycolor(input);
-    if (!color.isValid()) {
-      return null;
-    }
-    const obj = {};
-    const formats = ctx.herman.displayColors || ['hex', 'rgb', 'hsl'];
-    formats.forEach(format => {
-      let fn;
-      switch (format) {
-        case 'hex':
-          fn = 'toHexString';
-          break;
-        case 'rgb':
-          fn = 'toRgbString';
-          break;
-        case 'rgba':
-          format = 'rgb';
-          fn = 'toRgbString';
-          break;
-        case 'hsl':
-          fn = 'toHslString';
-          break;
-        case 'hsla':
-          format = 'hsl';
-          fn = 'toHslString';
-          break;
-      }
-      if (fn) {
-        obj[format] = color[fn]();
-      }
-    });
-    return obj;
-  });
-
-  dest = path.resolve(dest);
-
-  // set (external) base path for links
-  ctx.basePath = '';
-  ctx.activeProject = null;
-  ctx.topGroups = ctx.groups;
-  ctx.topByGroup = ctx.byGroup;
-
-  // check if we need to copy a favicon file or use the default
-  let copyShortcutIcon = false;
-  if (!ctx.shortcutIcon) {
-    ctx.shortcutIcon = { type: 'internal', url: 'assets/img/favicon.ico' };
-  } else if (ctx.shortcutIcon.type === 'internal') {
-    ctx.shortcutIcon.url = `assets/img/${ctx.shortcutIcon.url}`;
-    copyShortcutIcon = true;
-  }
-
-  // render the index template and copy the static assets.
-  const promises = [
-    render(nunjucksEnv, indexTemplate, indexDest, ctx),
-    copy(
-      path.join(assets, '/**/*.{css,js,ico,map}'),
-      path.join(dest, 'assets')
-    ).then(() => {
-      if (copyShortcutIcon) {
-        return copy(ctx.shortcutIcon.path, path.resolve(dest, 'assets/img/'));
-      }
-      return Promise.resolve();
-    }),
-  ];
-
-  if (ctx.customCSS) {
-    promises.push(
-      copy(ctx.customCSS.path, path.resolve(dest, 'assets/css/custom'), {
-        parser: parse.customCSS,
-        env: ctx,
-      }).then(() => {
-        if (ctx.customCSSFiles && ctx.customCSSFiles.length) {
-          return copy(ctx.customCSSFiles, path.resolve(dest, 'assets/custom'), {
-            base: ctx.dir,
-          });
-        }
-        return Promise.resolve();
-      })
-    );
-  }
-
-  if (ctx.localFonts && ctx.localFonts.length) {
-    promises.push(
-      copy(ctx.localFonts, path.resolve(dest, 'assets/fonts/'), {
-        base: path.resolve(ctx.dir, ctx.herman.fontpath),
-      })
-    );
-  }
-
-  const getRenderCtx = (context, groupName) =>
-    extend({}, context, {
-      pageTitle: context.groups[groupName],
-      activeGroup: groupName,
-      items: context.byGroup[groupName],
-    });
-
-  // Render a page for each group, too.
-  Object.getOwnPropertyNames(ctx.byGroup).forEach(groupName => {
-    const groupDest = path.join(dest, `${groupName}.html`);
-    const groupCtx = getRenderCtx(ctx, groupName);
-    promises.push(render(nunjucksEnv, groupTemplate, groupDest, groupCtx));
-  });
-
-  if (ctx.subprojects) {
-    // Render pages for subprojects.
-    Object.getOwnPropertyNames(ctx.subprojects).forEach(prjName => {
-      const prjCtx = ctx.subprojects[prjName];
-      const prjDest = path.join(dest, prjName);
-      const pageDest = path.join(prjDest, 'index.html');
-      promises.push(render(nunjucksEnv, indexTemplate, pageDest, prjCtx));
-
-      Object.getOwnPropertyNames(prjCtx.byGroup).forEach(groupName => {
-        const groupDest = path.join(prjDest, `${groupName}.html`);
-        const groupCtx = getRenderCtx(prjCtx, groupName);
-        promises.push(render(nunjucksEnv, groupTemplate, groupDest, groupCtx));
-      });
-    });
-  }
-
-  return Promise.all(promises);
-};
+const renderHerman = require('./lib/renderHerman');
 
 // get nunjucks env lazily so that we only throw an error on missing
 // templatepath if annotation was actually used.
-const getNunjucksEnv = (name, env, warned) => {
-  if (env.herman && env.herman.nunjucksEnv) {
-    return env.herman.nunjucksEnv;
-  }
-  if (!env.herman || !env.herman.templatepath) {
-    if (!warned) {
-      env.logger.warn(`Must pass in a templatepath if using ${name}.`);
-    }
-    return null;
-  }
-  return nunjucks.configure(env.herman.templatepath);
-};
+const getNunjucksEnv = require('./lib/getNunjucksEnv');
 
 /**
  * Actual theme function. It takes the destination directory `dest`,
