+++ lib/prepareContext.js
@@ -1,0 +1,177 @@
+module.exports = ctx => {
+  const def = {
+    display: {
+      access: ['public', 'private'],
+      alias: false,
+      watermark: true,
+    },
+    groups: {
+      undefined: 'General',
+    },
+    sort: ['group', 'file', 'line', 'access'],
+    herman: { sass: {} },
+  };
+
+  // Apply default values for groups and display.
+  ctx.groups = extend(def.groups, ctx.groups);
+  ctx.display = extend(def.display, ctx.display);
+
+  if (!ctx.description) {
+    def.descriptionPath = './README.md';
+  }
+
+  // Extend top-level context keys.
+  ctx = extend({}, def, ctx);
+
+  /**
+   * Load a `sass-json file` (if one is given in the context) and add its
+   * contents under the `sassjson` key of the context.
+   */
+  if (ctx.herman.sass && ctx.herman.sass.jsonfile && !ctx.sassjson) {
+    ctx.sassjson = parse.sassJson(fs.readFileSync(ctx.herman.sass.jsonfile));
+  }
+
+  /**
+   * Remove the bogus context from any standalone sassdoc comments.
+   * (We detect these if the context starts more than one line after the
+   * sassdoc comment ends.)
+   */
+  ctx.data.forEach(item => {
+    if (!item.context || !item.context.line) {
+      return;
+    }
+    if (item.context.type === 'unknown') {
+      item.context.type = 'prose';
+      item.context.line.end = item.context.line.start;
+    }
+    // Consider it to be prose if it's separated from the next Sass block
+    // by any blank lines.
+    const name = item.context.origName || item.context.name || '';
+    const lineCount = name.split('\n').length;
+    if (!name || item.context.line.start > item.commentRange.end + lineCount) {
+      item.context = {
+        type: 'prose',
+        line: item.commentRange,
+      };
+    }
+  });
+
+  /**
+   * Add `description` and `descriptionPath` from configuration.
+   * Note: `descriptionPath` will override `description`.
+   *
+   * See
+   * <http://sassdoc.com/extra-tools/#description-description-descriptionpath>.
+   */
+  extras.description(ctx);
+
+  /**
+   * Parse text data (like descriptions) as Markdown, and put the
+   * rendered HTML in `html*` variables.
+   *
+   * For example, `ctx.package.description` will be parsed as Markdown
+   * in `ctx.package.htmlDescription`.
+   *
+   * See <http://sassdoc.com/extra-tools/#markdown-markdown>.
+   */
+  extras.markdown(ctx);
+
+  /**
+   * Add a `display` property for each data item regarding of display
+   * configuration (hide private items and aliases for example).
+   *
+   * You'll need to add default values in your `.sassdocrc` before
+   * using this filter:
+   *
+   *     {
+   *       "display": {
+   *         "access": ["public", "private"],
+   *         "alias": false
+   *       }
+   *     }
+   *
+   * See <http://sassdoc.com/extra-tools/#display-toggle-display>.
+   */
+  extras.display(ctx);
+
+  /**
+   * Allow the user to give a name to the documentation groups.
+   *
+   * We can then have `@group slug` in the docblock, and map `slug`
+   * to `Some title string` in the theme configuration.
+   *
+   * **Note:** all items without a group are in the `undefined` group.
+   *
+   * See <http://sassdoc.com/extra-tools/#groups-aliases-groupname>.
+   */
+  extras.groupName(ctx);
+
+  /**
+   * Converts `shortcutIcon` config option into an object:
+   *
+   *       {
+   *         "type": "external",
+   *         "url": "http://absolute.path/to/icon.png"
+   *       }
+   *
+   *       {
+   *         "type": "internal",
+   *         "url": "icon.png",
+   *         "path": "/complete/relative/path/to/icon.png"
+   *       }
+   *
+   * See <http://sassdoc.com/extra-tools/#shortcut-icon-shortcuticon>.
+   */
+  extras.shortcutIcon(ctx);
+
+  /**
+   * Sorts items based on the `sort` config value.
+   * Sort order is determined by the last character: > (desc) or < (asc).
+   *
+   *       {
+   *         "sort": [
+   *            "access",
+   *            "line>",
+   *            "group",
+   *            "file"
+   *          ]
+   *       }
+   *
+   * See <http://sassdoc.com/extra-tools/#sort-sort>.
+   */
+  extras.sort(ctx);
+
+  /**
+   * Connects aliased variables to their original value.
+   * Adds `resolvedValue` key:
+   *
+   *       {
+   *         "description": "<p>Main color</p>\n",
+   *         "context": {
+   *           "type": "variable",
+   *           "name": "color-main",
+   *           "value": "$color-blue"
+   *         },
+   *         "type": "Color",
+   *         "resolvedValue": "#22b8dc"
+   *       }
+   *
+   * See <http://sassdoc.com/extra-tools/#resolved-variables-resolvevariables>.
+   */
+  extras.resolveVariables(ctx);
+
+  /**
+   * Index the data by group, so we have the following structure:
+   *
+   *     {
+   *       "group-slug": [...],
+   *       "another-group": [...]
+   *     }
+   *
+   * You can then use `data.byGroup` instead of `data` in your
+   * templates to manipulate the indexed object.
+   */
+  ctx.byGroup = byGroup(ctx.data, ctx.groups);
+
+  return ctx;
+};
