+++ lib/renderHerman.js
@@ -1,0 +1,130 @@
+module.exports = (dest, ctx) => {
+  const indexTemplate = path.join(base, 'index.j2');
+  const indexDest = path.join(dest, 'index.html');
+  const groupTemplate = path.join(base, 'group.j2');
+  const assets = path.resolve(__dirname, './dist');
+
+  // Accepts a color (in any format) and returns an object with hex, rgba, and
+  // hsla strings.
+  nunjucksEnv.addFilter('colors', input => {
+    const color = tinycolor(input);
+    if (!color.isValid()) {
+      return null;
+    }
+    const obj = {};
+    const formats = ctx.herman.displayColors || ['hex', 'rgb', 'hsl'];
+    formats.forEach(format => {
+      let fn;
+      switch (format) {
+        case 'hex':
+          fn = 'toHexString';
+          break;
+        case 'rgb':
+          fn = 'toRgbString';
+          break;
+        case 'rgba':
+          format = 'rgb';
+          fn = 'toRgbString';
+          break;
+        case 'hsl':
+          fn = 'toHslString';
+          break;
+        case 'hsla':
+          format = 'hsl';
+          fn = 'toHslString';
+          break;
+      }
+      if (fn) {
+        obj[format] = color[fn]();
+      }
+    });
+    return obj;
+  });
+
+  dest = path.resolve(dest);
+
+  // set (external) base path for links
+  ctx.basePath = '';
+  ctx.activeProject = null;
+  ctx.topGroups = ctx.groups;
+  ctx.topByGroup = ctx.byGroup;
+
+  // check if we need to copy a favicon file or use the default
+  let copyShortcutIcon = false;
+  if (!ctx.shortcutIcon) {
+    ctx.shortcutIcon = { type: 'internal', url: 'assets/img/favicon.ico' };
+  } else if (ctx.shortcutIcon.type === 'internal') {
+    ctx.shortcutIcon.url = `assets/img/${ctx.shortcutIcon.url}`;
+    copyShortcutIcon = true;
+  }
+
+  // render the index template and copy the static assets.
+  const promises = [
+    render(nunjucksEnv, indexTemplate, indexDest, ctx),
+    copy(
+      path.join(assets, '/**/*.{css,js,ico,map}'),
+      path.join(dest, 'assets')
+    ).then(() => {
+      if (copyShortcutIcon) {
+        return copy(ctx.shortcutIcon.path, path.resolve(dest, 'assets/img/'));
+      }
+      return Promise.resolve();
+    }),
+  ];
+
+  if (ctx.customCSS) {
+    promises.push(
+      copy(ctx.customCSS.path, path.resolve(dest, 'assets/css/custom'), {
+        parser: parse.customCSS,
+        env: ctx,
+      }).then(() => {
+        if (ctx.customCSSFiles && ctx.customCSSFiles.length) {
+          return copy(ctx.customCSSFiles, path.resolve(dest, 'assets/custom'), {
+            base: ctx.dir,
+          });
+        }
+        return Promise.resolve();
+      })
+    );
+  }
+
+  if (ctx.localFonts && ctx.localFonts.length) {
+    promises.push(
+      copy(ctx.localFonts, path.resolve(dest, 'assets/fonts/'), {
+        base: path.resolve(ctx.dir, ctx.herman.fontpath),
+      })
+    );
+  }
+
+  const getRenderCtx = (context, groupName) =>
+    extend({}, context, {
+      pageTitle: context.groups[groupName],
+      activeGroup: groupName,
+      items: context.byGroup[groupName],
+    });
+
+  // Render a page for each group, too.
+  Object.getOwnPropertyNames(ctx.byGroup).forEach(groupName => {
+    const groupDest = path.join(dest, `${groupName}.html`);
+    const groupCtx = getRenderCtx(ctx, groupName);
+    promises.push(render(nunjucksEnv, groupTemplate, groupDest, groupCtx));
+  });
+
+  if (ctx.subprojects) {
+    // Render pages for subprojects.
+    Object.getOwnPropertyNames(ctx.subprojects).forEach(prjName => {
+      const prjCtx = ctx.subprojects[prjName];
+      const prjDest = path.join(dest, prjName);
+      const pageDest = path.join(prjDest, 'index.html');
+      promises.push(render(nunjucksEnv, indexTemplate, pageDest, prjCtx));
+
+      Object.getOwnPropertyNames(prjCtx.byGroup).forEach(groupName => {
+        const groupDest = path.join(prjDest, `${groupName}.html`);
+        const groupCtx = getRenderCtx(prjCtx, groupName);
+        promises.push(render(nunjucksEnv, groupTemplate, groupDest, groupCtx));
+      });
+    });
+  }
+
+  return Promise.all(promises);
+};
